Add launchDebugger and --ld option to ktxapp
to make them available to all legacy apps.
