+++ tools/ktx2check/ktx2check.cpp
@@ -132,6 +132,9 @@ struct issue {
 };
 
 #define WARNING 0x00010000
+#if defined(ERROR) // windows.h defines this and is included by ktxapp.h.
+  #undef ERROR
+#endif
 #define ERROR 0x00100000
 #define FATAL 0x01000000
 
@@ -1104,46 +1107,6 @@ ktxValidator::usage()
     ktxApp::usage();
 }
 
-#if defined(DEBUG) && defined(NEED_TO_DEBUG_STDIN_WITH_VISUAL_STUDIO)
-#include <windows.h>
-
-bool launchDebugger()
-{
-    // Get System directory, typically c:\windows\system32
-    std::wstring systemDir(MAX_PATH + 1, '\0');
-    UINT nChars = GetSystemDirectoryW(&systemDir[0], systemDir.length());
-    if (nChars == 0) return false; // failed to get system directory
-    systemDir.resize(nChars);
-
-    // Get process ID and create the command line
-    DWORD pid = GetCurrentProcessId();
-    std::wostringstream s;
-    s << systemDir << L"\\vsjitdebugger.exe -p " << pid;
-    std::wstring cmdLine = s.str();
-
-    // Start debugger process
-    STARTUPINFOW si;
-    ZeroMemory(&si, sizeof(si));
-    si.cb = sizeof(si);
-
-    PROCESS_INFORMATION pi;
-    ZeroMemory(&pi, sizeof(pi));
-
-    if (!CreateProcessW(NULL, &cmdLine[0], NULL, NULL, FALSE, 0, NULL, NULL, &si, &pi)) return false;
-
-    // Close debugger process handles to eliminate resource leak
-    CloseHandle(pi.hThread);
-    CloseHandle(pi.hProcess);
-
-    // Wait for the debugger to attach
-    while (!IsDebuggerPresent()) Sleep(100);
-
-    // Stop execution so the debugger can take over
-    DebugBreak();
-    return true;
-}
-#endif
-
 int _tmain(int argc, _TCHAR* argv[])
 {
 
@@ -1159,10 +1122,6 @@ ktxValidator::main(int argc, _TCHAR *argv[])
 
     logger.quiet = options.quiet;
     logger.maxIssues = options.maxIssues;
-
-#if defined(DEBUG) && defined(NEED_TO_DEBUG_STDIN_WITH_VISUAL_STUDIO)
-    launchDebugger();
-#endif
 
     vector<_tstring>::const_iterator it;
     for (it = options.infiles.begin(); it < options.infiles.end(); it++) {
